+++ src/main/scala/streams/block/package.scala
@@ -8,7 +8,6 @@ import net.minecraft.util.math.BlockPos
 import net.minecraft.world._
 import net.minecraftforge.fml.relauncher.Side._
 import net.minecraftforge.fml.relauncher.SideOnly
-import scala.math.{atan2 => _}
 
 package object block {
 
@@ -17,8 +16,8 @@ package object block {
       state.getBlock match {
           case block: BlockRiver => // Fixed flow
               val flowVector = block.getFlow(w, pos, state)
-              if(flowVector.xCoord == 0D && flowVector.zCoord == 0D) -1000F
-              else (atan2(flowVector.zCoord, flowVector.xCoord) - PI / 2D).toFloat
+              if(flowVector.x == 0D && flowVector.z == 0D) -1000F
+              else (atan2(flowVector.z, flowVector.x) - PI / 2D).toFloat
           case _ => BlockLiquid.getSlopeAngle(w, pos, material, state)
       }
   }